+++ src/screens/retaliation/RetaliationCardViewModelMapper.ts
@@ -1,5 +1,5 @@
-import { RetaliationSocialIconImageMapper } from './../home/retaliation/RetaliationSocialIconImageMapper'
 import { Retaliation } from '../../core/entities/Retaliation'
+import { RetaliationSocialIconImageMapper } from '../retaliations/RetaliationSocialIconImageMapper'
 import { RetaliationCardViewModel } from './RetaliationCardViewModel'
 
 export const RetaliationCardViewModelMapper = {
